requirements.txt
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-semantic-conventions